requirements.txt
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
colorama==0.4.6
comtypes==1.4.16
cryptography==46.0.6
distro==1.9.0
dlib-bin==20.0.0
face-recognition==1.3.0
Flask==3.1.3
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
MarkupSafe==3.0.3
MouseInfo==0.1.3
numpy==1.26.4
openai==2.30.0
opencv-python==4.9.0.80
pillow==12.1.1
playsound==1.2.2
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyparsing==3.3.2
pyperclip==1.11.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
pyttsx3==2.99
pytweening==1.2.0
pytz==2026.1.post1
pywhatkit==5.4
pywin32==311
requests==2.33.0
sniffio==1.3.1
soupsieve==2.8.3
SpeechRecognition==3.15.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.7
wikipedia==1.4.0
gunicorn==21.2.0
firebase-admin==6.2.0
